phishing-urlextension/backend/requirements.txt
flask
scikit-learn
pandas
numpy
joblib
python-whois